codec: Improve codec::* struct state transitions (decoder)
This commit improves the state transitions for the `codec::*` structs as
discussed in https://github.com/meh/rust-ffmpeg/issues/7.
